// src/config/enums/config.enum.ts
export enum ConfigName {
	App = "app",
	Database = "database",
	Cloudinary = "cloudinary",
	Otp = "otp",
	Upload = "upload",
	Jwt = "jwt",
	Mailer = "mailer",
	UrlConfig = "urlConfig",
	RedisConfig = "redisConfig",
}

export enum NodeEnv {
	DEVELOPMENT = "DEVELOPMENT",
	PRODUCTION = "PRODUCTION",
}